[*] migration to ES and typescript, changed services and utils, configured ts
