+++ src/types/pino-multi-stream.d.ts
@@ -1,0 +1,5 @@
+declare module 'pino-multi-stream' {
+    import { StreamEntry, Logger } from 'pino';
+
+    export function multistream(streams: StreamEntry[]): Logger;
+}
